Add Sharing (Implicit intent)
Here we dont know or care how our request will be fulfilled. Eg Call a number

There are other types of implicit intent like
- Picking contacts from the  contacts content provider (content://contacts/people)
- Dailing a number
- Send an Email (https://www.codegrepper.com/code-examples/javascript/react+native+open+gmail+app)

Useful links
https://stackoverflow.com/questions/43214062/open-maps-google-maps-in-react-native

We dont really need to know which application that will fulfil this request and the user may not have the application installed in their app.